move css to ./styles
